pyproject.toml
[project]
name = "graphiti-cli"
version = "0.1.0"
description = "Command-line interface for Graphiti knowledge graph operations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "graphiti-core>=0.3.15",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
ai = [
    "claude-code-sdk>=0.0.14",
]

[project.scripts]
graphiti = "cli.graphiti_cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli"]
